responsive

+++ src/Controller/TrickController.php
@@ -11,6 +11,7 @@ use App\Form\TrickType;
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\EntityManagerInterface;
 use Embera\Embera;
+use PhpParser\Node\Expr\Cast\String_;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -27,11 +28,18 @@ class TrickController extends AbstractController
     public function __construct(SessionInterface $session){
         $this->session = $session;
     }
+
     /**
      * @Route("/ajouter", name="trickAdd")
      */
     public function addTrick(Request $request, EntityManagerInterface $manager): Response
     {
+        if(!$this->getUser()){
+            return $this->redirectToRoute('home');
+        }
+        if($this->getUser()->getActivationToken() != NULL){
+            return $this->redirectToRoute('home');
+        }
         $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
         $trick = new Trick();
         
@@ -41,6 +49,7 @@ class TrickController extends AbstractController
 
         if ($form->isSubmitted() && $form->isValid()) {
             $images = $form->get('images')->getData();
+            dump($images);
             foreach ($images as $image) {
                 $filename = md5(uniqid()) . '.' . $image->guessExtension();
 
@@ -55,6 +64,7 @@ class TrickController extends AbstractController
             }
 
             $videos = $form->get('videos')->getData();
+            dump($videos);
             foreach ($videos as $video) {
                 $url = preg_replace("/\s*[a-zA-Z\/\/:\.]*youtu(be.com\/watch\?v=|.be\/)([a-zA-Z0-9\-_]+)([a-zA-Z0-9\/\*\-\_\?\&\;\%\=\.]*)/i","<iframe src=\"//www.youtube.com/embed/$2\" allowfullscreen></iframe>",$video);
                 $vid = new Videos;
@@ -80,6 +90,12 @@ class TrickController extends AbstractController
      * @Route("/{id}/delete", name="trickDelete")
      */
     public function deleteTrick(string $id, EntityManagerInterface $entityManager){
+        if(!$this->getUser()){
+            return $this->redirectToRoute('home');
+        }
+        if($this->getUser()->getActivationToken() != NULL){
+            return $this->redirectToRoute('home');
+        }
         $trick = $this->getDoctrine()->getRepository(Trick::class)->find($id);
         $entityManager->remove($trick);
         $entityManager->flush();
@@ -88,7 +104,7 @@ class TrickController extends AbstractController
     }
 
     /**
-     * @Route("/{id}/show")
+     * @Route("/{id}/show", name="showTrick")
      */
     public function showTrick(string $id, Request $request, EntityManagerInterface $manager){
         $trick = $this->getDoctrine()->getRepository(Trick::class)->find($id);
@@ -107,5 +123,71 @@ class TrickController extends AbstractController
             'trick' => $trick,
             'commentForm' => $commentForm->createView()
         ]);
+    }
+
+    /**
+     * @Route("/{id}/update", name="updateTrick")
+     */
+    public function updateTrick(Trick $trick, Request $request){
+        if(!$this->getUser()){
+            return $this->redirectToRoute('home');
+        }
+        if($this->getUser()->getActivationToken() != NULL){
+            return $this->redirectToRoute('home');
+        }
+        $form = $this->createForm(TrickType::class, $trick);
+        $form->handleRequest($request);
+
+        if($form->isSubmitted() && $form->isValid()){
+            $images = $form->get('images')->getData();
+            foreach ($images as $image) {
+                $filename = md5(uniqid()) . '.' . $image->guessExtension();
+
+                $image->move(
+                    $this->getParameter('images_directory'),
+                    $filename
+                );
+
+                $img = new Images();
+                $img->setFilename($filename);
+                $trick->addImage($img);
+            }
+            $videos = $form->get('videos')->getData();
+            foreach ($videos as $video) {
+                dump($videos);
+                $url = preg_replace("/\s*[a-zA-Z\/\/:\.]*youtu(be.com\/watch\?v=|.be\/)([a-zA-Z0-9\-_]+)([a-zA-Z0-9\/\*\-\_\?\&\;\%\=\.]*)/i","<iframe src=\"//www.youtube.com/embed/$2\" allowfullscreen></iframe>", $video);
+                $vid = new Videos;
+                $vid->setUrl($url);
+                $trick->addVideo($vid);
+            }
+
+            $this->getDoctrine()->getManager()->flush();
+
+            $this->session->getFlashBag()->add('message', 'Votre trick à bien été modifier !');
+            return $this->redirectToRoute('updateTrick', ['id' => $trick->getId()]);
+        }
+        return $this->render('pages/updateTrick.html.twig', [
+            'trick' => $trick,
+            'form' => $form->createView()
+        ]);
+    }
+
+    /**
+     * @Route("/{idTrick}/images/{idImage}/delete")
+     */
+    public function deleteImage(string $idTrick, string $idImage){
+        if(!$this->getUser()){
+            return $this->redirectToRoute('home');
+        }
+        if($this->getUser()->getActivationToken() != NULL){
+            return $this->redirectToRoute('home');
+        }
+
+        $image = $this->getDoctrine()->getRepository(Images::class)->find($idImage);
+        $manager = $this->getDoctrine()->getManager();
+        $manager->remove($image);
+        $manager->flush();
+
+        return $this->redirect("/tricks/$idTrick/update");
     }
 }
